[feat] Implement "add book" functionality
Coded the submitClick function which will query the HTML form for book
information and send it to the addBookToLibrary() function. This will
add books to the library array.

Also added the invalidEntryWarning() function, which will be a
placeholder until something better can be implemented in the future.

Needed to change the HTML form from a hard coded <form> to straight
input boxes with no <form> tag. Was too difficult to handle the
onsubmit event without reloading the page. Now submitClick() will just
clear fields (need to refactor into its own function).