[feat] Implement localStorage
On load, the script checks that localStorage is available, then if a
library array exists. If no, it creates one. Else, it loads the library
array. If localStorage is not available, it alerts the user that the
library will not be saved.

Every time a book is added, modified, or deleted, the localStorage is
updated.